fix : db 타입 변경

+++ backend/game-service/src/main/java/yare/yare/domain/price_option/entity/Season.java
@@ -20,7 +20,7 @@ public class Season {
     private Integer id;
 
     @NotNull
-    @Column(columnDefinition = "TINYINT")
+    @Column(columnDefinition = "SMALLINT")
     private Integer year;
 
     @NotNull
